+++ lib/features/home/cubit/statistics_state.dart
@@ -4,7 +4,7 @@ part of 'statistics_cubit.dart';
 class StatisticsState with _$StatisticsState {
   const factory StatisticsState.initial() = _Initial;
   const factory StatisticsState.empty() = _Empty;
-  const factory StatisticsState.notEmpty({required int amountExpense, required int amountIncome}) = _NotEmpty;
-  const factory StatisticsState.uploadStatistics({required int amountExpense, required int amountIncome}) = _UploadStatistics;
+  const factory StatisticsState.notEmpty({required double amountExpense, required double amountIncome}) = _NotEmpty;
+  const factory StatisticsState.uploadStatistics({required double amountExpense, required double amountIncome}) = _UploadStatistics;
 
 }
