+++ lib/features/new_transaction/cubit/total_state.dart
@@ -3,5 +3,5 @@ part of 'total_cubit.dart';
 @freezed
 class TotalState with _$TotalState {
   const factory TotalState.initial() = _Initial;
-  const factory TotalState.changeTotal({required int total}) = _ChangeTotal;
+  const factory TotalState.changeTotal({required double total}) = _ChangeTotal;
 }
